rename default getstaticpaths
