+++ main-site/lib/getStatic/default.getStaticPaths.ts
