+++ main-site/pages/[slug]/tag/[tagId].celebTagPage.tsx
@@ -100,5 +100,5 @@ export default function TagPage(props: TagPageProps) {
   );
 }
 
-export { getStaticPaths } from '~/lib/getStatic/defaultGetStaticPaths';
+export { getStaticPaths } from '~/lib/getStatic/default.getStaticPaths';
 export { getStaticProps } from '~/lib/getStatic/celebTagPage.getStaticProps';
