+++ main-site/pages/~issue/[issueId].issuePage.tsx
@@ -8,5 +8,5 @@ export default function IssuePage(props: any) {
   );
 }
 
-export { getStaticPaths } from '~/lib/getStatic/defaultGetStaticPaths';
+export { getStaticPaths } from '~/lib/getStatic/default.getStaticPaths';
 export { getStaticProps } from '~/lib/getStatic/issuePage.getStaticProps';
